// packages/activities-service/src/persistence/activities/acitivities.types.ts
export type PersistenceActivity = {
  id: number;
  has_streams: boolean;
  name: string;
  distance: number;
  distance_miles: number;
  moving_time: number;
  elapsed_time: number;
  total_elevation_gain: number;
  type: string;
  average_seconds_per_mile: number;
  min_seconds_per_mile: number;
  max_seconds_per_mile: number;
  sport_type: string;
  start_date: Date;
  start_date_local: Date;
  timezone: string;
  utc_offset: number;
  location_city: string;
  location_state: string;
  location_country: string;
  achievement_count: number;
  kudos_count: number;
  comment_count: number;
  athlete_count: number;
  photo_count: number;
  trainer: boolean;
  commute: boolean;
  manual: boolean;
  private: boolean;
  visibility: string;
  flagged: boolean;
  gear_id: string;
  start_latlng: { coordinates: number[][]; } | null;
  end_latlng: { coordinates: number[][]; } | null;
  average_speed: number;
  max_speed: number;
  has_heartrate: boolean;
  average_heartrate: number;
  max_heartrate: number;
  heartrate_opt_out: boolean;
  display_hide_heartrate_option: boolean;
  elev_high: number;
  elev_low: number;
  upload_id: number;
  upload_id_str: string;
  external_id: string;
  from_accepted_tag: boolean;
  pr_count: number;
  total_photo_count: number;
  has_kudoed: boolean;
  summary_polyline: string;
  hidden: boolean;
};

export type PersistenceBestEffort = {
  effort_id: number;
  start_date_local: Date;
  distance: number;
  elapsed_time: number;
  moving_time: number;
  pr_rank: number;
  name: string;
  start_index: number;
  end_index: number;
};

export type PersistenceCalculatedBestEffort = {
  start_date_local: Date;
  distance: number;
  elapsed_time: number;
  moving_time: number;
  pr_rank?: number;
  name: string;
  start_index: number;
  end_index: number;
  activityId: number;
};

export type PersistenceRelatedActivity = {
  baseActivity: number;
  relatedActivity: number;
  segmentScoreFromBase: number | null;
  segmentScoreFromRelated: number | null;
  longestCommonSegmentSubsequence: number | null;
  numberBaseSegments: number | null;
  numberRelatedSegments: number | null;
  routeScoreFromBase: number | null;
  routeScoreFromRelated: number | null;
};